Now Licenses does all the job when converting to_s.
It has two possibilities: if there's an license that's Copyleft or Share Alike,
we consider that this licenses group'll have to be licensed under this license.
So, when printing, we just print it (this might have a bug, because we consider
that Share Alike licenses are compatible with newer versions of the same
license. So, if there is BY-SA 2.0 and BY-SA 3.0, we'll just print the first we
find (and not the later version)).

If there are only Lesser Copyleft and Permissive licenses, we print their name
+ version sepparated by commas and the grouping of their attributes.